Remove base_object from Ruby model
and move methods from base_object to each model so that model files will
have no dependency when generating "models only"
