+++ modules/swagger-codegen/src/main/resources/ruby/gem.mustache
@@ -5,7 +5,6 @@ require '{{gemName}}/version'
 require '{{gemName}}/configuration'
 
 # Models
-require '{{gemName}}/{{modelPackage}}/base_object'
 {{#models}}
 require '{{importPath}}'
 {{/models}}
